Add text addressing citable code as requested from editor
This is a first go at this, please look at and revise.  For instance we
could add more about other ways that we do this.  There is also an
interesting discussion not yet discussed as to the likelihood that
GitHub is a long-term solution (as in decades) for the scientific
literature.