Make it easy to download notebooks from the Sphinx-rendered version.
One downside of nbsphinx is that it's less obvious how to get
a copy of the notebook (compared to if we used, say, nbviewer or
Github to render them).  You can click on 'show source' to get
the notebook file, but the extension '.txt' is appended, which
will cause problems for people who don't know how to fix it
(and annoy everyone).  This commit ensures that "show source"
on a rendered notebook links to the notebook file with the
correct (.ipynb) extension.

docathon