Make PyClaw gallery work with multiple examples in each directory.
This was previously written under the assumption that there
was only one example in each subdirectory of pclaw/examples.
Nowadays there are multiple examples in some directories
(different problems but with the same set of equations).
This updates the gallery-building scripts to work with that
situation by using e.g. "_plots_advection_1d" instead of just "_plots".

docathon